fix: address review feedback for manifest cache changes
